Refactor and improve for A21 release
Add grass texture resources for reliable patching

Load original atlas from our own resource, as it was impossible to get the original
Mip (best quality) reliably. It works when the game loads first, but when another
game is started from the main menu, the resulting textures are always blurry. Tried
really hard to avoid to embed to original textures, but that was the easiest fix.
Alternative was to hold on to initial textures, but that would waste 50MB of RAM.